+++ apps.py
@@ -9,6 +9,6 @@ from utils import init_models
 DICT_MODELS = init_models()
 logger = init_graylogger()
 
-path_to_config = os.getenv("APP_SETTINGS", "rest_flask.config.Config")
+path_to_config = os.getenv("APP_SETTINGS", "config.Config")
 CONFIG = import_string(path_to_config)
-LOGGER = Logger(CONFIG.APP_NAME)
+LOGGER = Logger(CONFIG.SERVICE_NAME)