+++ mq_rabbit/process/__init__.py
@@ -1,0 +1,2 @@
+from .command_lgbm import CommandLgbm
+from .command_random import CommandRandom