+++ utils/__init__.py
@@ -1,1 +1,2 @@
 from .model_handler import init_models
+from .rabbit import Rabbit