+++ CHANGES.md
@@ -21,3 +21,9 @@
 - (2021-10-28) Merge to develop
 
 - (2021-10-28) Added base structure rest_flask
+
+- (2021-10-28) Merge to develop
+
+- (2021-11-01) Added base structure mq_rabbit
+
+- (2021-10-28) Merge to develop